fix(db-codegen,inspector): address CodeRabbit review findings
- db-codegen: fix PK columns being emitted as nullable (`!c.notNull && !c.pk`)
- custom-types-generator: fix `findMatchingAngleBracket` treating `>` in `=>` as a generic closer
- schema-generator: handle async tsx `unregister()` rejections instead of fire-and-forget
- EmailsPage: extract `EmailsOverview` into its own file with a named props type
